Add command-line arguments
This commit adds command-line arguments to MarketDB. MarketDB is now
invoked like so:

  marketdb ticker

For example,

  marketdb NASDAQ:MSFT

This will retrieve Microsoft Corporation's data from the database
and provide a quote for that security. As part of this, a function -
get_id_from_ticker - was created which find the security's ID provided
its ticker.